Added - dropdown, settings(TBC), Working splash,
splash is now 'home' and there is quick access to 'homepage' for now until login system fully functional. dropdown added for settings and logout, which takes you back to the splash page. made sidebar and header content blocked so that can be overridden for splash page. no links working yet for splash except temporary 'homepage' and "ENTER" to login